requirements.txt
# Local Tailor — Requirements
# ============================
# Install: pip install -r requirements.txt
#
# First run downloads (cached in ~/.cache/huggingface after first run):
#   all-MiniLM-L6-v2                 ~80MB   (SetFit base model)
#   deepset/roberta-base-squad2      ~500MB  (span extraction)

# Core NLP
setfit>=1.0.0
sentence-transformers>=2.2.0
transformers>=4.30.0,<5.0.0  # setfit 1.x incompatible with transformers 5.x
datasets>=2.12.0
torch>=2.0.0          # CPU-only is sufficient; GPU optional

# Data
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0

# UI
streamlit>=1.28.0

# Report generation
jinja2>=3.1.0
fpdf2>=2.7.0          # PDF reports; pure Python, no system deps

# Utilities
tqdm>=4.65.0
